+++ internal/provider/fwprovider/provider.go
@@ -28,6 +28,7 @@ import (
 	sysconfig "github.com/hashicorp/terraform-provider-vault/internal/vault/sys/config"
 
 	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
+	"github.com/hashicorp/terraform-plugin-framework/action"
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
 	"github.com/hashicorp/terraform-plugin-framework/provider"
 	"github.com/hashicorp/terraform-plugin-framework/provider/schema"
@@ -35,10 +36,12 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-provider-vault/internal/consts"
 	sdkv2provider "github.com/hashicorp/terraform-provider-vault/internal/provider"
+	"github.com/hashicorp/terraform-provider-vault/internal/vault/secrets/rotate"
 	"github.com/hashicorp/terraform-provider-vault/internal/vault/sys"
 )
 
 var _ provider.ProviderWithEphemeralResources = &fwprovider{}
+var _ provider.ProviderWithActions = &fwprovider{}
 
 // Ensure the implementation satisfies the provider.Provider interface
 var _ provider.Provider = &fwprovider{}
@@ -229,6 +232,7 @@ func (p *fwprovider) Configure(ctx context.Context, req provider.ConfigureReques
 	resp.DataSourceData = v
 	resp.ResourceData = v
 	resp.EphemeralResourceData = v
+	resp.ActionData = v
 }
 
 // Resources returns a slice of functions to instantiate each Resource
@@ -319,5 +323,16 @@ func (p *fwprovider) DataSources(ctx context.Context) []func() datasource.DataSo
 		pki_external_ca.NewPKIExternalCAOrderChallengeDataSource,
 		sys.NewPluginRuntimesDataSource,
 		config.NewSysConfigCORSDataSource,
+	}
+}
+
+// Actions returns a slice of functions to instantiate each Action
+// implementation.
+//
+// The action type name is determined by the Action implementing
+// the Metadata method. All actions must have unique names.
+func (p *fwprovider) Actions(_ context.Context) []func() action.Action {
+	return []func() action.Action{
+		rotate.NewRotateRootAction,
 	}
 }